Add person logic extracted to API Business Logic
